+++ toposens_driver/doc/html/search/all_1.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['getframe',['getFrame',['../classtoposens__driver_1_1Serial.html#a7e1ba7bccecbc978f6e16104ec459e29',1,'toposens_driver::Serial']]]
-];