+++ toposens_driver/doc/html/search/all_4.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['poll',['poll',['../classtoposens__driver_1_1Sensor.html#a926de2fe6169b81fb456d2a8d726cebe',1,'toposens_driver::Sensor']]]
-];