+++ toposens_driver/doc/html/search/classes_0.js
@@ -1,5 +1,0 @@
-var searchData=
-[
-  ['sensor',['Sensor',['../classtoposens__driver_1_1Sensor.html',1,'toposens_driver']]],
-  ['serial',['Serial',['../classtoposens__driver_1_1Serial.html',1,'toposens_driver']]]
-];