+++ toposens_driver/doc/html/search/files_0.js
@@ -1,6 +1,0 @@
-var searchData=
-[
-  ['sensor_2eh',['sensor.h',['../sensor_8h.html',1,'']]],
-  ['serial_2ecpp',['serial.cpp',['../serial_8cpp.html',1,'']]],
-  ['serial_2eh',['serial.h',['../serial_8h.html',1,'']]]
-];