+++ toposens_driver/doc/html/search/functions_4.js
@@ -1,7 +1,0 @@
-var searchData=
-[
-  ['send',['send',['../classtoposens__driver_1_1Serial.html#aea7a45dbb1d2edb0b5e60b1cae10584b',1,'toposens_driver::Serial']]],
-  ['sensor',['Sensor',['../classtoposens__driver_1_1Sensor.html#ab14040539bf943fcfb95d613ace04f80',1,'toposens_driver::Sensor']]],
-  ['serial',['Serial',['../classtoposens__driver_1_1Serial.html#ab66b3b0aba714ac8332d7453833f4294',1,'toposens_driver::Serial']]],
-  ['shutdown',['shutdown',['../classtoposens__driver_1_1Sensor.html#a1a568d4c7946731b20c3f81a3c67a566',1,'toposens_driver::Sensor']]]
-];