+++ toposens_driver/doc/html/search/functions_5.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['_7eserial',['~Serial',['../classtoposens__driver_1_1Serial.html#a7970e34fd2b626baa989a7a08e3b2e7f',1,'toposens_driver::Serial']]]
-];