+++ toposens_driver/doc/html/search/variables_1.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['kbaud',['kBaud',['../classtoposens__driver_1_1Serial.html#a383754d17418659020d9f7c221457f5f',1,'toposens_driver::Serial']]]
-];