+++ toposens_driver/doc/html/search/variables_2.js
@@ -1,4 +1,0 @@
-var searchData=
-[
-  ['kportid',['kPortId',['../classtoposens__driver_1_1Serial.html#aed7003fc9c3014a5f565ff49b63a62d8',1,'toposens_driver::Serial']]]
-];