+++ toposens_driver/src/ts_driver_node.cpp
@@ -7,31 +7,29 @@
 // Signal-safe flag for whether shutdown is requested
 volatile sig_atomic_t g_shutdown = 0;
 
-// Replacement SIGINT handler
+// Replacement SIGINT handler. Called on Ctrl-C
 void onSigint(int sig)
 {
   g_shutdown = 1;
 }
 
-// Replacement "shutdown" XMLRPC callback
+// Replacement "shutdown" XMLRPC callback. Called on rosnode kill.
 void onShutdown(XmlRpc::XmlRpcValue& params, XmlRpc::XmlRpcValue& result)
 {
   int num_params = 0;
-  if (params.getType() == XmlRpc::XmlRpcValue::TypeArray)
-    num_params = params.size();
+  if (params.getType() == XmlRpc::XmlRpcValue::TypeArray) num_params = params.size();
   if (num_params > 1)
   {
     std::string reason = params[1];
     ROS_WARN("Shutdown request received. Reason: [%s]", reason.c_str());
     g_shutdown = 1; // Set flag
   }
-
   result = ros::xmlrpc::responseInt(1, "", 0);
 }
 
 
 int main(int argc, char** argv)
-{ 
+{
   ros::init(argc, argv, "ts_driver_node", ros::init_options::NoSigintHandler);
   signal(SIGINT, onSigint); // Override SIGINT handler
 
@@ -57,7 +55,7 @@ int main(int argc, char** argv)
   } catch (const char *msg) {
     ROS_ERROR("%s", msg);
   }
-
+  
   ros::shutdown();
   return 0;
 }