+++ toposens_markers/src/ts_markers_node.cpp
@@ -1,6 +1,6 @@
 #include <ros/ros.h>
 
-#include "toposens_markers/marker.h"
+#include "toposens_markers/markers.h"
 
 
 int main(int argc, char** argv)
@@ -9,7 +9,7 @@ int main(int argc, char** argv)
   ros::NodeHandle nh;
   ros::NodeHandle private_nh("~");
 
-  toposens_markers::Marker m(nh, private_nh);
+  toposens_markers::Markers m(nh, private_nh);
   ros::spin();
 
   return 0;